+++ services/director-v2/src/simcore_service_director_v2/modules/osparc_variables/substitutions.py
@@ -29,7 +29,7 @@ from ...utils.osparc_variables import (
     resolve_variables_from_context,
 )
 from ..db.repositories.services_environments import ServicesEnvironmentsRepository
-from ._user_session import request_user_email, request_user_role
+from ._user import request_user_email, request_user_role
 from .api_keys_manager import get_or_create_user_api_key, get_or_create_user_api_secret
 
 _logger = logging.getLogger(__name__)
